+++ suuchi-core/src/main/scala/in/ashwanthkumar/suuchi/partitioner/PartitionNode.scala
@@ -1,6 +1,6 @@
 package in.ashwanthkumar.suuchi.partitioner
 
-import in.ashwanthkumar.suuchi.membership.MemberAddress
+import in.ashwanthkumar.suuchi.cluster.MemberAddress
 
 case class VNode(node: MemberAddress, nodeReplicaId: Int) {
   def key = node.host+"_"+node.port+"_"+nodeReplicaId
