+++ suuchi-core/src/main/scala/in/ashwanthkumar/suuchi/partitioner/Partitioner.scala
@@ -1,6 +1,6 @@
 package in.ashwanthkumar.suuchi.partitioner
 
-import in.ashwanthkumar.suuchi.membership.MemberAddress
+import in.ashwanthkumar.suuchi.cluster.MemberAddress
 
 import scala.util.hashing.MurmurHash3
