+++ suuchi-core/src/main/scala/in/ashwanthkumar/suuchi/router/Marshallers.scala
@@ -1,6 +1,6 @@
 package in.ashwanthkumar.suuchi.router
 
-import in.ashwanthkumar.suuchi.membership.MemberAddress
+import in.ashwanthkumar.suuchi.cluster.MemberAddress
 import io.grpc.Metadata.AsciiMarshaller
 
 /**
