+++ suuchi-core/src/main/scala/in/ashwanthkumar/suuchi/router/ReplicationRouter.scala
@@ -3,7 +3,7 @@ package in.ashwanthkumar.suuchi.router
 import java.util.concurrent.{Executor, Executors}
 
 import com.google.common.util.concurrent.{Futures, ListenableFuture}
-import in.ashwanthkumar.suuchi.membership.MemberAddress
+import in.ashwanthkumar.suuchi.cluster.MemberAddress
 import in.ashwanthkumar.suuchi.rpc.CachedChannelPool
 import io.grpc.ServerCall.Listener
 import io.grpc._
