+++ suuchi-core/src/test/scala/in/ashwanthkumar/suuchi/router/HandleOrForwardRouterSpec.scala
@@ -1,6 +1,6 @@
 package in.ashwanthkumar.suuchi.router
 
-import in.ashwanthkumar.suuchi.membership.MemberAddress
+import in.ashwanthkumar.suuchi.cluster.MemberAddress
 import io.grpc.ServerCall.Listener
 import io.grpc._
 import org.mockito.Matchers._
