+++ suuchi-core/src/test/scala/in/ashwanthkumar/suuchi/router/ListOfNodesMarshallerSpec.scala
@@ -1,6 +1,6 @@
 package in.ashwanthkumar.suuchi.router
 
-import in.ashwanthkumar.suuchi.membership.MemberAddress
+import in.ashwanthkumar.suuchi.cluster.MemberAddress
 import org.scalatest.FlatSpec
 import org.scalatest.Matchers.{be, contain, convertToAnyShouldWrapper, have}
