+++ suuchi-core/src/test/scala/in/ashwanthkumar/suuchi/router/SequentialReplicatorSpec.scala
@@ -1,6 +1,6 @@
 package in.ashwanthkumar.suuchi.router
 
-import in.ashwanthkumar.suuchi.membership.MemberAddress
+import in.ashwanthkumar.suuchi.cluster.MemberAddress
 import io.grpc.{Metadata, MethodDescriptor, ServerCall, Status}
 import org.mockito.Matchers
 import org.mockito.Matchers._
